separate ensureContextMenu
