+++ src/background/messenger/strict/api.ts
@@ -123,3 +123,5 @@ export const collectPerformanceDiagnostics = getMethod(
 
 export const initTelemetry = getNotifier("INIT_TELEMETRY", bg);
 export const sendDeploymentAlert = getNotifier("SEND_DEPLOYMENT_ALERT", bg);
+
+export const ensureContextMenu = getMethod("ENSURE_CONTEXT_MENU", bg);
